rename targetselection.py to decals.py
Paving out the way for applying TS to mocks.

Likely we will need a copy of decals.py for TS of mocks; some duplication of
code is needed, but we could able to refactor them away later on, when we have
the code.

I will try to do an example for the Lightcone input. but please send me a test file.