Determines file order for mock data reads
Previously this was left to the implicit order from os.walk in
desitarget.io.iter_files. Now the ordering of the data from each file in
the concatenated output is set by the result of np.argsort applied to
the input file names (names only, not full paths).

Useful for associating a predictable integer index with input files for
each mock without relying on coding that into the individual mock file
names.