Adds standard star generation from mocks
Adds a new bin script, make_standards.py

Standards are read from mocks in the same way as targets, but they get
special targetids. OBSCONDITIONS for standards are a hack right now,
this needs to be done properly.

Introduces a reserved range of MTL targetids for skies and standards
that builds on previous assumptions for how make_mtl re-writes
targetids. Making use of this is implemented for standards in this
commit, but not for skies.